refactor: Renamed function (engine)
