+++ src/migration/engine/steps/create-file-path-maps/addon/map-component-classes.ts
