+++ src/migration/engine/steps/create-file-path-maps/tests/map-components.ts
