+++ tests/migration/engine/steps/create-file-path-maps/typescript.test.ts
@@ -1,13 +1,13 @@
 import { assert, loadFixture, test } from '@codemod-utils/tests';
 
-import { createFilePathMaps } from '../../../../../src/migration/ember-engine/steps/index.js';
+import { createFilePathMaps } from '../../../../../src/migration/engine/steps/index.js';
 import { inputProject } from '../../../../fixtures/ember-engine/typescript/index.js';
 import {
   codemodOptions,
   options,
 } from '../../../../helpers/shared-test-setups/ember-engine/typescript.js';
 
-test('migration | ember-engine | steps | create-file-path-maps > typescript', function () {
+test('migration | engine | steps | create-file-path-maps > typescript', function () {
   loadFixture(inputProject, codemodOptions);
 
   const filePathMaps = createFilePathMaps(options);
