+++ tests/migration/engine/steps/create-options/javascript.test.ts
@@ -1,13 +1,13 @@
 import { assert, loadFixture, test } from '@codemod-utils/tests';
 
-import { createOptions } from '../../../../../src/migration/ember-engine/steps/index.js';
+import { createOptions } from '../../../../../src/migration/engine/steps/index.js';
 import { inputProject } from '../../../../fixtures/ember-engine/javascript/index.js';
 import {
   codemodOptions,
   options,
 } from '../../../../helpers/shared-test-setups/ember-engine/javascript.js';
 
-test('migration | ember-engine | steps | create-options > javascript', function () {
+test('migration | engine | steps | create-options > javascript', function () {
   loadFixture(inputProject, codemodOptions);
 
   assert.deepStrictEqual(createOptions(codemodOptions), options);
