+++ tests/migration/engine/steps/create-options/relative-imports.test.ts
@@ -1,13 +1,13 @@
 import { assert, loadFixture, test } from '@codemod-utils/tests';
 
-import { createOptions } from '../../../../../src/migration/ember-engine/steps/index.js';
+import { createOptions } from '../../../../../src/migration/engine/steps/index.js';
 import { inputProject } from '../../../../fixtures/ember-engine/relative-imports/index.js';
 import {
   codemodOptions,
   options,
 } from '../../../../helpers/shared-test-setups/ember-engine/relative-imports.js';
 
-test('migration | ember-engine | steps | create-options > relative-imports', function () {
+test('migration | engine | steps | create-options > relative-imports', function () {
   loadFixture(inputProject, codemodOptions);
 
   assert.deepStrictEqual(createOptions(codemodOptions), options);
